pyproject.toml
[tool.poetry]
name = "prediction-markets"
version = "0.1.0"
description = "Prediction market forecasting with LLM baselines and custom agent"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.28"
openai = "^1.60"
anthropic = "^0.45"
google-genai = "^1.5"
pandas = "^2.2"
pyarrow = "^18.0"
matplotlib = "^3.9"
seaborn = "^0.13"
pydantic = "^2.10"
pydantic-settings = "^2.7"
pyyaml = "^6.0"
tavily-python = "^0.5"
tenacity = "^9.0"
tqdm = "^4.67"
rich = "^13.9"
python-dateutil = "^2.9"
python-dotenv = "^1.0"
streamlit = "^1.55.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-asyncio = "^0.25"
ruff = "^0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
